# docs/openapi-frontend-separation.yaml
openapi: 3.0.3
info:
  title: fictionweb 前后端分离接口清单
  version: 1.0.0
  description: |
    面向前后端分离改造的接口清单（OpenAPI 风格）。
    - 统一响应：ResponseData，成功 code=200
    - 统一鉴权：请求头 token（RSA 公钥加密后的字符串）
    - WebSocket 日志流：/api/log/ws（不属于标准 OpenAPI HTTP 接口，见文末说明）
servers:
  - url: /
...
security:
  - tokenAuth: []
paths:
...
  /api/queryBookList:
    post:
      tags: [书籍]
      summary: 查询书籍列表
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookSearchVO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBookPage'
  /api/modelBook/{id}:
    get:
      tags: [书籍]
      summary: 查询单个书籍
      parameters:
        - $ref: '#/components/parameters/IdPathLong'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBook'
  /api/saveBook:
    post:
      tags: [书籍]
      summary: 保存书籍（bookid 为空新增，否则修改）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseVoid'
  /api/deleteBook/{id}:
    get:
      tags: [书籍]
      summary: 删除书籍（多个 id 用逗号分隔）
      parameters:
        - $ref: '#/components/parameters/IdPathString'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseVoid'
...
  /api/downloadBook/{id}:
    get:
      tags: [书籍]
      summary: 下载书籍（方式1）
      parameters:
        - $ref: '#/components/parameters/IdPathLong'
      responses:
        '200':
          description: 文件流
          content:
            application/octet-stream: {}
  /api/downloadBook2/{id}:
    get:
      tags: [书籍]
      summary: 下载书籍（方式2）
      parameters:
        - $ref: '#/components/parameters/IdPathLong'
      responses:
        '200':
          description: 文件流
          content:
            application/octet-stream: {}
...
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: token
      description: RSA 公钥加密后的 token 字符串
  parameters:
    IdPathLong:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    IdPathString:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 单个 id 或逗号分隔 id 列表
  schemas:
    ResponseBase:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: ""
      required: [code, message]
    ResponseVoid:
      allOf:
        - $ref: '#/components/schemas/ResponseBase'
        - type: object
          properties:
            data:
              nullable: true
...
    ResponseBook:
      allOf:
        - $ref: '#/components/schemas/ResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Book'
...
    ResponseBookPage:
      allOf:
        - $ref: '#/components/schemas/ResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageListBookListReturnVO'
...
    PageOrder:
      type: object
      properties:
        pageNum:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          minimum: 1
          example: 10
        orderBy:
          type: string
        descOrAsc:
          type: string
          enum: [Desc, Asc]
...
    BookSearchVO:
      allOf:
        - $ref: '#/components/schemas/PageOrder'
        - type: object
          properties:
            name:
              type: string
            siteid:
              type: integer
              format: int64
              nullable: true
...
    PageListBase:
      type: object
      properties:
        total:
          type: integer
          format: int64
        pageNum:
          type: integer
        pageSize:
          type: integer
        pages:
          type: integer
        hasPreviousPage:
          type: boolean
        hasNextPage:
          type: boolean
...
    PageListBookListReturnVO:
      allOf:
        - $ref: '#/components/schemas/PageListBase'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/BookListReturnVO'
...
    Book:
      type: object
      properties:
        bookid:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        url:
          type: string
        configid:
          type: integer
          format: int64
        starturl:
          type: string
        handlerinfo:
          type: string
          description: JSON 字符串
        siteid:
          type: integer
          format: int64
        taskstatus:
          type: integer
          description: 0无，1重新获取全部，2获取后续章节，3重新获取单章
        taskswitch:
          type: integer
          description: 0关闭，1开启
        detail:
          type: string
        lastgettime:
          type: string
          format: date-time
        createtime:
          type: string
          format: date-time
    BookListReturnVO:
      allOf:
        - $ref: '#/components/schemas/Book'
        - type: object
          properties:
            configname:
              type: string
            sitename:
              type: string
            taskstatusname:
              type: string
            taskswitchname:
              type: string
...